Move `NullLifecycle` into a Sub-Namespace
Because we'll add a few more classes to core to handle some composition.
